baler: img2 query - ts, comp_id wrong round off fix
In bhttpd img2 query, ts_end is calculated by
  ts_end = ts_begin + npp*width
which is mathematically correct. However, the expression is of float
type, and got round-off to int as ts_end is of type int. By losing the
precision, sometimes the result is wrong (e.g. ts_end < ts_begin).

This commit fixes the bug by casting the addition part to int:
  ts_end = ts_begin + (int)(npp*width)
The addition part should not exceed the precision of the mantissa.
Hence, we yield the correct result.

The host_end calculation also follow the same logic above.

+++ baler/src/bhttpd/bhttpd_query.c
@@ -902,8 +902,8 @@ void bhttpd_handle_query_img2(struct bhttpd_req_ctxt *ctxt)
 	int height = __get_url_param_int(ctxt, "height", 0);
 	int ts_begin = __get_url_param_int(ctxt, "ts_begin", 0);
 	int host_begin = __get_url_param_int(ctxt, "host_begin", 0);
-	int ts_end = ts_begin + width*spp;
-	int host_end = host_begin + height*npp;
+	int ts_end = ts_begin + (int)(width*spp);
+	int host_end = host_begin + (int)(height*npp);
 
 	int *data = NULL;
